+++ src/spaces/jolt_job_system.hpp
@@ -4,11 +4,13 @@ class JoltJobSystem final : public JPH::JobSystemWithBarrier {
 public:
 	JoltJobSystem();
 
-	static JoltJobSystem* get_singleton() { return singleton; }
-
 	void pre_step();
 
 	void post_step();
+
+#ifdef GDJ_CONFIG_EDITOR
+	void flush_timings();
+#endif // GDJ_CONFIG_EDITOR
 
 private:
 	class Job : public JPH::JobSystem::Job {
@@ -40,6 +42,10 @@ private:
 
 		inline static std::atomic<Job*> completed_head = nullptr;
 
+#ifdef GDJ_CONFIG_EDITOR
+		const char* name = nullptr;
+#endif // GDJ_CONFIG_EDITOR
+
 		int64_t task_id = -1;
 
 		std::atomic<Job*> completed_next = nullptr;
@@ -60,7 +66,13 @@ private:
 
 	void FreeJob(JPH::JobSystem::Job* p_job) override;
 
-	inline static JoltJobSystem* singleton = nullptr;
+#ifdef GDJ_CONFIG_EDITOR
+	// HACK(mihe): We use `const void*` here to avoid the cost of hashing the actual string, since
+	// the job names are always literals and as such will point to the same address every time.
+	inline static HashMap<const void*, uint64_t> timings_by_job;
+
+	inline static SpinLock timings_lock;
+#endif // GDJ_CONFIG_EDITOR
 
 	FreeList<Job> jobs;